fix(typecheck): restore green root CI - DOM-free ChatEvent + TS6 BufferSource
The root `bun x tsc --noEmit` (CI's typecheck) regressed when P-COLLAB.1
landed: harness/scripts/demo_pcollab1.ts imports desktop/collab/crypto.ts ->
frames.ts, which imported `ChatEvent` from renderer/bridge.ts - dragging bridge.ts
(a DOM file: window/document/fetch) into the root program, whose lib is
["ESNext"] only. That surfaced ~10 "Cannot find name 'document'/'window'/
'HTMLMetaElement'" + "Property 'data' does not exist" errors, plus two
crypto.ts WebCrypto errors under TS 6's stricter `BufferSource`
(`Uint8Array<ArrayBufferLike>` no longer matches `Uint8Array<ArrayBuffer>`), and
a noUncheckedIndexedAccess slip in the demo. Master has been red since #268; I'd
only been checking desktop/tsconfig.json (which has the DOM lib) locally.

Fix, keeping the root program DOM-free (correct for a node/harness typecheck):
- extract the pure `ChatEvent` union into desktop/renderer/chat_events.ts (its
  only non-primitive members are DOM-free harness types); bridge.ts re-exports it
  unchanged so every existing importer is untouched.
- point desktop/collab frames.ts + host.ts at chat_events.ts (not bridge.ts).
- crypto.ts `strict()` now returns `Uint8Array<ArrayBuffer>` so every WebCrypto
  call site satisfies BufferSource.
- demo_pcollab1.ts: guard the tamper byte for noUncheckedIndexedAccess.

Root tsc + desktop tsc both 0 errors; 29 collab tests + both demos + license green.
